Support THL (Taiwan Huayu Luomapinyin) and MPS2 Pinyin keyboard
THL (Taiwan Huayu Luomapinyin)[*] is derived from Hanyu Pinyin system and
adapted according to the conventions in Taiwan.  This patch introduces
two key layouts: THL and MPS2 Pinyin (注音二式).  Also, the case when
one key sequence is mapped to several Pinyin representations is
considered.  The new-added phoneAlt in struct ZuinData is used for such
purpose.  For example:

  KB_THL_PINYIN: (Taiwan Huayu Luomapinyin)
  phone: 0x1c80  (ㄒㄧ)
  phoneAlt: 0x2a00 (ㄙ)

  KB_HANYU_PINYIN:
  phone: 0x2a00 (ㄙ)
  phoneAlt: 0x1c80  (ㄒㄧ)

The phonetic representation will be decided upon the given keyboard layout.

Check the following discussion for details:
  https://groups.google.com/group/chewing-devel/browse_thread/thread/4b5aefc87356d2fe

[*] http://thl.tw/about

Signed-off-by: Jim Huang <jserv@0xlab.org>