Fix ChoiceEndChoice expression always true
The chiSymbolCursor should be placed back to old position no matter
which type of choice selection is finished.
The original expression wants to check the type of choice, then set
 the chiSymbolCursor to old position, but the expression is always
true. The chiSymbolCursor should place back to old position after
finishing the choice selection in any case, so removing the if
statement.