Update chewing_handle_CtrlNum()
1. NoSymbolBetween() and copyStringFromPreeditBuf() don't skip any
non-chinese characters, they require the cursor of pre-edit buffer.
So replace the PhoneSeqCursor with chiSymbolCursor.

2. When users press ctrl-2 ... ctrl-9, but there's nothing to be
added (for example, the pre-edit buffer is empty, or there are some
symbols in between), libchewing will display the old aux message,
or just nothing. However this is not desirable, it's better to
display an error message instead.

3. This function expects keys from 0 to 9.
If not, return -1 at the end.

4. add or update test cases.